Fix encoding of command strings sent to the einhorn master
url.QueryEscape encodes spaces to '+', which doesn't work because
Ruby's URI.decode doesn't decode them. It also wasn't technically
consistent with the interface anyway, since only % and \n were
supposed to be escaped.

Instead of trying to use generic functions to do the escaping, just
replace % and \n with their percent-encoded counterparts in the
string. It's not overly robust, but it also doesn't need to be.