Adds support for PROXY protocol
This adds support for the PROXY protocol, the original client's IP to be
tunneled through to smokescreen without editing the original request.
This can be useful when putting a TCP load balancer (eg. a Classic ELB)
in front of the service.

This support needs to be enabled via a new option: -proxy-protocol

The underlying library tries to do the right thing by examining the
initial bytes in the request. If the PROXY protocol isn't being used,
the request will still be processed.

The connection's address will be reported based on what the PROXY
protocol passed, so only requests from trusted upstreams should be
allowed (otherwise malicious systems could spoof their IP).