Reference req rather than ctx.Req in handler
`req` is what goproxy will eventually use to make the actual request
so I think it makes more sense to treat that as canonical. I believe it
will currently be the same as ctx.Req but it would be possible in the
future for someone to add another handler that changes it.