Log complete TCP tuples where possible
We log information about TCP connections accepted by the proxy or
created by the proxy, but we don't do it consistently.  We also use
terms like "src" and "dst", which are ambiguous in the context of two
TCP connections (the inbound connection from a client and the outbound
connection that we're making on behalf of the client).

This commit changes how we log TCP information, preferring terms like
"inbound," "outbound," "local," and "remote" to label the various pieces
of the IP+TCP 4-tuples.  It also adds the address and port number for
the local side of outgoing connections, to aid in debugging network
issues.  Finally, since we're changing these field names anyway, we will
prefer "addr" over "ip".