[Fix] Add validation for Default ACL Rules (#248)
The default rule within an ACL configuration file was not subjected to the same validation checks as named service-specific rules. Specifically, if an administrator globally disables an enforcement policy (e.g., the "open" policy), this disablement is not honored for the default rule.

This change checks that the every rule in ACL including default rule has a conformant domain glob and is not utilizing a disabled enforcement policy